requirements.txt
Flask
Flask-Cors
requests==2.31.0
python-dotenv==1.0.0
google-generativeai
looker-sdk==23.16.0
gunicorn==21.2.0